+++ src/components/Select/SuggestionsList/SuggestionsList.tsx
@@ -1,8 +1,10 @@
 import * as React from 'react';
 import { GetItemPropsOptions } from 'downshift';
 import { bem } from '../../../utils/bem';
+import { ContentPlaceholder } from '../../ContentPlaceholder';
 import { Text, MarkedText } from '../../Text';
 import { ListOptimizer, ListItem } from '../../List';
+import { NUMBER_OF_SUGGESTION_LOADING_PLACEHOLDERS } from '../../../constants';
 import styles from './SuggestionsList.scss';
 
 const { elem } = bem('SuggestionsList', styles);
@@ -10,12 +12,20 @@ const { elem } = bem('SuggestionsList', styles);
 export interface Props<S> {
     /** An array of objects that will be used to render the suggestions list. */
     suggestions: S[];
+    /** if suggestions are still loading, i.e. display placeholders */
+    isLoading?: boolean;
     /** Enable ListOptimizer component for decreasing render time */
     useOptimizeRender?: boolean;
     /** suggestionToString(suggestion) should return a string to be displayed in the UI. e.g.: suggestion => suggestion.name */
-    suggestionToString: (suggestions: S) => string;
+    suggestionToString: (suggestion: S) => string;
+    /** suggestionToKey(suggestion) makes a key to be used for a suggestion item */
+    suggestionToKey?: (suggestion: S) => string;
+    /** render function for suggestion list item */
+    suggestionItemRenderer?: (suggestion: S) => ReactNode;
     /** to be shown when no suggestions are available */
     noSuggestionsPlaceholder?: string;
+    /** Defines if the first item of suggestions list is always visible */
+    isFirstItemAlwaysVisible?: boolean;
     /** a function which gets props for the item in the list */
     getItemProps: (options: GetItemPropsOptions<S>) => object;
     /** index of the item from the list to be highlighted */
@@ -29,22 +39,15 @@ export function SuggestionsList<S>(props: Props<S>) {
         suggestionToString,
         useOptimizeRender,
         suggestions,
+        isLoading,
+        isFirstItemAlwaysVisible,
         noSuggestionsPlaceholder,
         getItemProps,
         highlightedIndex,
+        suggestionToKey,
+        suggestionItemRenderer,
         inputValue,
     } = props;
-
-    if (!suggestions || !suggestions.length) {
-        if (noSuggestionsPlaceholder) {
-            return (
-                <ListItem disabled>
-                    <Text context="muted">{noSuggestionsPlaceholder}</Text>
-                </ListItem>
-            );
-        }
-        return null;
-    }
 
     // eslint-disable-next-line react/display-name
     const renderItem = ({ key, index, style = {} }) => (
@@ -58,11 +61,49 @@ export function SuggestionsList<S>(props: Props<S>) {
             isHighlighted={highlightedIndex === index}
             highlightContext="brand"
         >
-            <MarkedText marker={inputValue} inline>
-                {suggestionToString(suggestions[index])}
-            </MarkedText>
+            {suggestionItemRenderer ? (
+                suggestionItemRenderer(suggestions[index])
+            ) : (
+                <MarkedText marker={inputValue} inline>
+                    {suggestionToString(suggestions[index])}
+                </MarkedText>
+            )}
         </ListItem>
     );
+
+    // eslint-disable-next-line react/display-name
+    const renderLoadingPlaceholders = () =>
+        Array(NUMBER_OF_SUGGESTION_LOADING_PLACEHOLDERS)
+            .fill('')
+            .map((el, i) => {
+                if (isFirstItemAlwaysVisible && i === 0) {
+                    return renderItem({ key: 'firstItem', index: 0 });
+                }
+
+                return (
+                    // eslint-disable-next-line react/no-array-index-key
+                    <ListItem key={i}>
+                        <div {...elem('loaderItem')}>
+                            <ContentPlaceholder />
+                        </div>
+                    </ListItem>
+                );
+            });
+
+    if (isLoading) {
+        return <>{renderLoadingPlaceholders()}</>;
+    }
+
+    if (!suggestions || !suggestions.length) {
+        if (noSuggestionsPlaceholder) {
+            return (
+                <ListItem disabled>
+                    <Text context="muted">{noSuggestionsPlaceholder}</Text>
+                </ListItem>
+            );
+        }
+        return null;
+    }
 
     return (
         <>
@@ -73,7 +114,9 @@ export function SuggestionsList<S>(props: Props<S>) {
             ) : (
                 <>
                     {suggestions.map((item, index) => {
-                        const key = suggestionToString(item);
+                        const key = suggestionToKey
+                            ? suggestionToKey(item)
+                            : suggestionToString(item);
                         return renderItem({ key, index });
                     })}
                 </>
@@ -85,6 +128,8 @@ export function SuggestionsList<S>(props: Props<S>) {
 SuggestionsList.defaultProps = {
     noSuggestionsPlaceholder: '',
     useOptimizeRender: false,
+    isLoading: false,
+    isFirstItemAlwaysVisible: false,
     inputValue: '',
 };