+++ stories/Select.tsx
@@ -45,6 +45,10 @@ storiesOf('Organisms|Select', module)
             store.set({ selectedSuggestions: [...store.get('selectedSuggestions'), item] });
         };
 
+        const onFocus = () => {
+            console.log('onFocus was called');
+        };
+
         const onBlur = () => {
             console.log('onBlur was called');
             setTimeout(() => store.set({ inputValue: '' }), 200);
@@ -63,10 +67,12 @@ storiesOf('Organisms|Select', module)
                         'Use optimize list render (ListOptimizer)',
                         false
                     )}
+                    isLoading={boolean('isLoading', false)}
                     inputPlaceholder={text('Input placeholder', 'Select something...')}
                     noSuggestionsPlaceholder={text('No suggestions', 'No suggestions found...')}
                     suggestions={getSuggestions()}
                     suggestionToString={SUGGESTION_TO_STRING}
+                    onFocus={onFocus}
                     onBlur={onBlur}
                     onSelectionChange={onSelectionChange}
                     onInputValueChange={onInputValueChange}
@@ -80,16 +86,20 @@ storiesOf('Organisms|Select', module)
         searchFor.name = `Search for "${store.get('inputValue')}"`;
         searchFor.value = store.get('inputValue');
         const getSuggestions = (): TSuggestion[] => {
-            if (!store.get('inputValue').length) return [];
             const suggestions = SUGGESTIONS.filter(
                 (item: TSuggestion) =>
+                    item.name.toLocaleLowerCase() !== store.get('inputValue').toLocaleLowerCase() &&
+                    item.name
+                        .toLocaleLowerCase()
+                        .includes(store.get('inputValue').toLocaleLowerCase()) &&
                     !store
                         .get('selectedSuggestions')
-                        .some((sug: TSuggestion) => sug.name === item.name)
-            ).filter((item: TSuggestion) =>
-                item.name.toLocaleLowerCase().includes(store.get('inputValue').toLocaleLowerCase())
+                        .some((i) => item.name.toLocaleLowerCase() === i.name.toLocaleLowerCase())
             );
-            return [searchFor, ...suggestions];
+            if (store.get('inputValue').length) {
+                return [searchFor, ...suggestions];
+            }
+            return suggestions;
         };
 
         const onInputValueChange = (value: string) => {
@@ -118,6 +128,7 @@ storiesOf('Organisms|Select', module)
 
         const onSelectionRemove = (item: TSuggestion) => {
             console.log(`onSelectionRemove was called with {name: ${item.name}}`);
+            store.set({ inputValue: '' });
             const selectedItem = { ...item };
             // Delete item
             if (!store.get('inputValue')) {
@@ -140,6 +151,14 @@ storiesOf('Organisms|Select', module)
             store.set({ selectedSuggestions: [] });
         };
 
+        const onFocus = () => {
+            console.log('onFocus was called');
+        };
+
+        const onSubmit = () => {
+            console.log('onSubmit was called');
+        };
+
         return (
             <div style={{ width: '500px' }}>
                 <AutosuggestMulti
@@ -149,9 +168,16 @@ storiesOf('Organisms|Select', module)
                     suggestions={getSuggestions()}
                     suggestionToString={SUGGESTION_TO_STRING}
                     onBlur={onBlur}
+                    onFocus={onFocus}
+                    onSubmit={onSubmit}
+                    isFirstItemAlwaysVisible={
+                        boolean('First item is always visible', false) &&
+                        !!store.get('inputValue').length
+                    }
                     onSelectionChange={onSelectionChange}
                     onSelectionRemove={onSelectionRemove}
                     isProminent={boolean('Use prominent styling', true)}
+                    isLoading={boolean('isLoading', false)}
                     onInputValueChange={onInputValueChange}
                     showClearButton={boolean('Show clear button', true)}
                     clearTitle={text('Clear button label', 'Clear')}